20150218.0955

+++ cmdMULTIsens/lib/TMP36/TMP36.cpp
@@ -36,7 +36,7 @@ float TMP36::readTemperature(void)
     delay(110);
   }
   // measurement = (float)sumSamples / (float)cntSamples; // Calculate avg raw value.
-  measurement = (float)sumSamples * _invsamples; // Calculate avg raw value.
+  measurement = (float)sumSamples * (float)_invsamples; // Calculate avg raw value.
   measurement = map(measurement * 10, 0, 10230, 0, _ref5v * 10000) * 0.0001;
   return 100 * measurement - 50.0;
 }
